+++ tfdm/tfdm_main.cpp
@@ -420,7 +420,7 @@ static void parseCommandline(int32_t argc, const char* argv[]) {
                     hpprintf("Invalid option.\n");
                     exit(EXIT_FAILURE);
                 }
-                *ori = qRotateZ(atof(argv[i + 1]) * pi_v<float> / 180) * *ori;
+                *ori = qRotateZ(static_cast<float>(atof(argv[i + 1])) * pi_v<float> / 180) * *ori;
                 i += 1;
             }
             else if (strncmp(arg, "-pitch", 7) == 0) {
@@ -428,7 +428,7 @@ static void parseCommandline(int32_t argc, const char* argv[]) {
                     hpprintf("Invalid option.\n");
                     exit(EXIT_FAILURE);
                 }
-                *ori = qRotateX(atof(argv[i + 1]) * pi_v<float> / 180) * *ori;
+                *ori = qRotateX(static_cast<float>(atof(argv[i + 1])) * pi_v<float> / 180) * *ori;
                 i += 1;
             }
             else if (strncmp(arg, "-yaw", 5) == 0) {
@@ -436,7 +436,7 @@ static void parseCommandline(int32_t argc, const char* argv[]) {
                     hpprintf("Invalid option.\n");
                     exit(EXIT_FAILURE);
                 }
-                *ori = qRotateY(atof(argv[i + 1]) * pi_v<float> / 180) * *ori;
+                *ori = qRotateY(static_cast<float>(atof(argv[i + 1])) * pi_v<float> / 180) * *ori;
                 i += 1;
             }
         };
